remake some tests and methods of classes

+++ solving/CSR_Matrix/csr_matrix.h
@@ -13,7 +13,7 @@ struct Indexes{
 };
 
 struct CSRMatrixData{
-    vector<double>  values;
+    vector<float>  values;
     vector<size_t> columns_indexes;
     vector<int> number_rows_non_0_elements;
 };
@@ -24,18 +24,18 @@ bool operator<(const Indexes &first_indexes, const Indexes &second_indexes);
 
 class CSRMatrix {
 public:
-    CSRMatrix(const map<Indexes,double> &matrix);
+    CSRMatrix(const map<Indexes,float> &matrix);
     float GiveElement(const size_t &number_row, const size_t &number_column) const;
     vector<float> MultiplyByColumn(const vector<float> &column);
     CSRMatrixData GiveInformation() const;
     friend bool operator==(const CSRMatrix &first_matrix, const CSRMatrix &second_matrix);
 
 private:
-    vector<double>  values_;
+    vector<float>  values_;
     vector<size_t> columns_indexes_;
     vector<int> number_rows_non_0_elements_;
 
-    vector<int>CountNonZeroRowElements(const map<Indexes,double> &matrix);
+    vector<int>CountNonZeroRowElements(const map<Indexes,float> &matrix);
     float FindElement(const size_t &number_row, const size_t &number_column) const;
 
     vector<float> MultiplicationColumn(const vector<float> &column);
